+++ src/main/resources/db/migration/V2__Create_question_table.sql
@@ -1,12 +1,12 @@
 create table question
 (
-	id int auto_increment,
+	id bigint auto_increment,
 	title varchar(50),
 	description text,
 	tag varchar(256),
 	gem_create bigint,
 	gem_modified bigint,
-	creator int,
+	creator bigint,
 	comment_count int default 0,
 	view_count int default 0,
 	like_count int default 0,
